+++ App/app/layout.tsx
@@ -7,8 +7,8 @@ import Footer from '@/components/Footer'
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Upstat',
+  description: 'The Upstat Project',
 }
 
 export default function RootLayout({
