+++ App/app/page.module.css
@@ -1,0 +1,9 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  gap: 100px;
+  align-items: center;
+  padding: 6rem;
+  min-height: 100vh;
+}
+
