+++ App/package.json
@@ -12,6 +12,8 @@
     "@types/node": "20.8.2",
     "@types/react": "18.2.24",
     "@types/react-dom": "18.2.8",
+    "eslint": "8.50.0",
+    "eslint-config-next": "13.5.4",
     "next": "13.5.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
