--- veem/app/app.style.ts
@@ -1,8 +1,0 @@
-import { styled } from "styled-components";
-
-
-const HomeContainer = styled.div`
-    
-`
-
-export { HomeContainer }
