--- veem/app/dashboard/dashboard.style.ts
@@ -1,9 +1,0 @@
-"use client"
-import { styled } from "styled-components";
-
-
-const DashboardContainer = styled.div`
-    
-`
-
-export { DashboardContainer }
