--- veem/components/Footer.tsx
@@ -1,7 +1,0 @@
-const Footer = () => (
-    <footer>
-        <h1>Footer goes here</h1>
-    </footer>
-)
-
-export default Footer
