--- veem/components/Header.tsx
@@ -1,7 +1,0 @@
-const Header = () => (
-    <header>
-        <h1>Header goes here</h1>
-    </header>
-)
-
-export default Header
