--- veem/next.config.js
@@ -1,6 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-    pageExtensions: ['page.tsx', 'page.ts', 'page.jsx', 'page.js']
-}
-
-module.exports = nextConfig
